Fix a bug where jagged arrays with null values in the overhang
The code that calculates the maximum length of value in each column of
the Whitespace SV file had the max line length calculation in the wrong
block. This meant that the max_line_len array would not contain
calculations for overhanging null values. For example, parsing and
writing the following Whitespace SV file with right or left column
alignment would cause a panic:

1
2 -